requirements.txt
discord.py==2.3.0
i18nice
python-dotenv